Lots of changes, some world loading, improved EncryptedSocketInputStream::read Also added -Wall to CXXFLAGS, which should have been there from the start.

+++ src/game/block/Block.cpp
@@ -3,7 +3,7 @@
 
 namespace MCServer {
 
-Block::Block(uint16_t id, uint8_t metadata)
+Block::Block(uint8_t id, uint8_t metadata)
 :id(id), metadata(metadata) {
     
 }
